labs/10-analyze-text/python/text-analysis/requirements.txt
azure-ai-textanalytics
python-dotenv